extracting function to load locks by network
